remove circular dependency
