+++ .pytest_cache/v/cache/lastfailed
@@ -1,0 +1,1 @@
+{}
