+++ .pytest_cache/v/cache/nodeids
@@ -1,0 +1,32 @@
+[
+  "fda/basis.py::fda.basis.BSpline",
+  "fda/basis.py::fda.basis.FDataBasis",
+  "fda/basis.py::fda.basis.FDataBasis.from_data",
+  "fda/basis.py::fda.basis.FDataBasis.mean",
+  "fda/basis.py::fda.basis.FDataBasis.to_grid",
+  "fda/basis.py::fda.basis.Fourier",
+  "fda/basis.py::fda.basis.Monomial",
+  "fda/basis.py::fda.basis.Monomial.penalty",
+  "fda/grid.py::fda.grid.FDataGrid",
+  "fda/grid.py::fda.grid.FDataGrid.concatenate",
+  "fda/grid.py::fda.grid.FDataGrid.derivative",
+  "fda/grid.py::fda.grid.FDataGrid.to_basis",
+  "fda/kernel_smoothers.py::fda.kernel_smoothers.knn",
+  "fda/kernel_smoothers.py::fda.kernel_smoothers.local_linear_regression",
+  "fda/kernel_smoothers.py::fda.kernel_smoothers.nw",
+  "fda/math.py::fda.math.inner_product",
+  "fda/math.py::fda.math.metric",
+  "fda/math.py::fda.math.norm_lp",
+  "fda/validation.py::fda.validation.minimise",
+  "tests/test_basis.py::TestBasis::test_bspline_penalty",
+  "tests/test_basis.py::TestBasis::test_bspline_penalty_numerical",
+  "tests/test_basis.py::TestBasis::test_bspline_penalty_special_case",
+  "tests/test_basis.py::TestBasis::test_fourier_penalty",
+  "tests/test_basis.py::TestBasis::test_from_data_cholesky",
+  "tests/test_basis.py::TestBasis::test_from_data_qr",
+  "tests/test_basis.py::TestBasis::test_monomial_smoothing",
+  "tests/test_grid.py::TestFDataGrid::test_gmean",
+  "tests/test_grid.py::TestFDataGrid::test_init",
+  "tests/test_grid.py::TestFDataGrid::test_mean",
+  "tests/test_grid.py::TestFDataGrid::test_slice"
+]
